Move function getVerseListBookNumber from AppController to verseListController
